+++ packages/ast-spec/tests/util/serializers/string.ts
@@ -1,4 +1,4 @@
-import { NewPlugin } from 'pretty-format';
+import type { NewPlugin } from 'pretty-format';
 
 // custom string serializer so that we can use single-quoted strings instead of double quoted strings
 // this plays nicer with the way that the snapshot diff result, which is a pure string