+++ packages/eslint-plugin/src/rules/consistent-type-assertions.ts
@@ -1,5 +1,7 @@
+import type { TSESTree } from '@typescript-eslint/utils';
+import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+
 import * as util from '../util';
-import { TSESTree, AST_NODE_TYPES } from '@typescript-eslint/utils';
 
 // intentionally mirroring the options
 type MessageIds =