+++ packages/type-utils/src/propertyTypes.ts
@@ -1,4 +1,4 @@
-import * as ts from 'typescript';
+import type * as ts from 'typescript';
 
 export function getTypeOfPropertyOfName(
   checker: ts.TypeChecker,