+++ packages/typescript-estree/tests/lib/parse.test.ts
@@ -1,12 +1,48 @@
 import debug from 'debug';
 import { join, resolve } from 'path';
+
 import * as parser from '../../src';
-import * as astConverter from '../../src/ast-converter';
-import { TSESTreeOptions } from '../../src/parser-options';
-import * as sharedParserUtils from '../../src/create-program/shared';
+import * as astConverterModule from '../../src/ast-converter';
+import * as sharedParserUtilsModule from '../../src/create-program/shared';
+import type { TSESTreeOptions } from '../../src/parser-options';
 import { createSnapshotTestBlock } from '../../tools/test-utils';
 
 const FIXTURES_DIR = join(__dirname, '../fixtures/simpleProject');
+
+// we can't spy on the exports of an ES module - so we instead have to mock the entire module
+jest.mock('../../src/ast-converter', () => {
+  const astConverterActual = jest.requireActual<typeof astConverterModule>(
+    '../../src/ast-converter',
+  );
+
+  return {
+    ...astConverterActual,
+    __esModule: true,
+    astConverter: jest.fn(astConverterActual.astConverter),
+  };
+});
+jest.mock('../../src/create-program/shared', () => {
+  const sharedActual = jest.requireActual<typeof sharedParserUtilsModule>(
+    '../../src/create-program/shared',
+  );
+
+  return {
+    ...sharedActual,
+    __esModule: true,
+    createDefaultCompilerOptionsFromExtra: jest.fn(
+      sharedActual.createDefaultCompilerOptionsFromExtra,
+    ),
+  };
+});
+
+const astConverterMock = jest.mocked(astConverterModule.astConverter);
+const createDefaultCompilerOptionsFromExtra = jest.mocked(
+  sharedParserUtilsModule.createDefaultCompilerOptionsFromExtra,
+);
+
+beforeEach(() => {
+  jest.clearAllMocks();
+});
 
 describe('parseWithNodeMaps()', () => {
   describe('basic functionality', () => {
@@ -114,8 +150,6 @@ describe('parseWithNodeMaps()', () => {
 
   describe('loggerFn should be propagated to ast-converter', () => {
     it('output tokens, comments, locs, and ranges when called with those options', () => {
-      const spy = jest.spyOn(astConverter, 'astConverter');
-
       const loggerFn = jest.fn(() => {});
 
       parser.parseWithNodeMaps('let foo = bar;', {
@@ -126,8 +160,8 @@ describe('parseWithNodeMaps()', () => {
         loc: true,
       });
 
-      expect(spy).toHaveBeenCalled();
-      expect(spy.mock.calls[0][1]).toMatchObject({
+      expect(astConverterMock).toHaveBeenCalled();
+      expect(astConverterMock.mock.calls[0][1]).toMatchObject({
         code: 'let foo = bar;',
         comment: true,
         comments: [],
@@ -625,16 +659,11 @@ describe('parseAndGenerateServices', () => {
     });
 
     it('should turn on typescript debugger', () => {
-      const spy = jest.spyOn(
-        sharedParserUtils,
-        'createDefaultCompilerOptionsFromExtra',
-      );
-
       parser.parseAndGenerateServices('const x = 1;', {
         debugLevel: ['typescript'],
       });
-      expect(spy).toHaveBeenCalled();
-      expect(spy).toHaveReturnedWith(
+      expect(createDefaultCompilerOptionsFromExtra).toHaveBeenCalled();
+      expect(createDefaultCompilerOptionsFromExtra).toHaveReturnedWith(
         expect.objectContaining({
           extendedDiagnostics: true,
         }),