+++ packages/website/docusaurusConfig.ts
@@ -1,14 +1,14 @@
 import type { MDXPlugin } from '@docusaurus/mdx-loader';
 import type { Options as PluginContentDocsOptions } from '@docusaurus/plugin-content-docs';
 import type { Options as PresetClassicOptions } from '@docusaurus/preset-classic';
+import npm2yarnPlugin from '@docusaurus/remark-plugin-npm2yarn';
 import type { UserThemeConfig as ThemeCommonConfig } from '@docusaurus/theme-common';
 import type { UserThemeConfig as AlgoliaThemeConfig } from '@docusaurus/theme-search-algolia';
 import type { Config } from '@docusaurus/types';
-
-import { rulesMeta } from './rulesMeta';
-import npm2yarnPlugin from '@docusaurus/remark-plugin-npm2yarn';
 import tabsPlugin from 'remark-docusaurus-tabs';
+
 import { generatedRuleDocs } from './plugins/generated-rule-docs';
+import { rulesMeta } from './rulesMeta';
 
 const remarkPlugins: MDXPlugin[] = [[npm2yarnPlugin, { sync: true }]];